Refactored and found issues involving globals.
Showed program to dad. He was impressed, but pointed out a few instances
of globals that should be removed. Namely the spawnAsteroid method and the
player's moveSelf method.

For refactoring we removed the child variable from all object child
classes and replaced it with the python built-in type method. This gives
the same result but removes the need for the additional variable and
strange name that it had.

He suggests that all of the classes inheriting from the object parent
class should be given a reference to the game class so that they can
refer to it in functions.
Alternatively, I could pass in the required variables for each function
when it is called. However this could affect modularity if many different
versions of a funcion are passed redundant variables.