Removed instances of globals.
Classes that inherit from the Object class now are given a refernce to
the game class in it's super.()__init__() method.

However, in the asteroidSpawn method it is given a direct reference to
game in its first argument.
I don't really like this solution and I need to find a better way to
pass a reference to the asteroid and its variables into that function.
However, this may be very difficult, and it may also be simply better
pass game into this function as one of the arguments.

To Dos:
Add proper art
Add sound effects
Done (probably)