Added new graphics, fuel and highscore file
I added new pixel art graphics and a fuel mechanic. At the start of the
game the player is given 10 seconds worth of fuel. If this fuel runs
out, they lose. They can replenish their fuel supply by picking up jerry
cans (previously coins). This increases their score by one and allows
then to survive for longer. Highscores are now kept in a json file that
is read at the start of the game and written to when the game is closed.
This allows for persistent highscores across game sessions.

Next I am hoping to add: the player sprite rotating to face the
direction it is travelling and a fuel bar at the top of the screen that
roughly shows the amount of fuel a player has left. I also want to add a
highscore viewer to allow the player to see the top ~10 scores.