Ship now faces direction its moving in
The player ship now faces the direction that it is moving in. It does
this by using trigonometry: tan(x) = speedY / speedX and then solving
for x. The image can then be rotated x anticlockwise degrees, where 0
degrees means facing towards the right hand side.

Next I am hoping to add a highscore option to the menu which opens a
screen that shows the top ~10 scores.