Added highscore viewer, improved README
I have added a menu option to allow a user to view the top 20
highscores. The highscores are read from the highscores.json file and I
use python string formatting to ensure consistent spacing.

I have also improved the README to include a screenshot and commands
that can be easily copied and pasted into the terminal.