Added animations
I have added animations for when the player dies. There is an explosion
animation that plays when the player collides with an asteroid and if
the player runs out of fuel, an animation plays where smoke leaks from
the exhaust tubes. The frames of each animation are exported as
individual .png files, rather than a sprite sheet, which is easier for me
at the moment.