Added credits screen
I have added a credits screen that can be accessed from the main menu.
Currently, it credits me for the art, code, music and sound as well as
Stefie Justprince for the original pixelify font, which I have modified
and used for the game. I may add more acreditations later.