Removed unneeded files
I have removed unneeded files from the repository, but kept local copies
on my machine. This includes sound effects and images that are no longer
used and an incomplete option to train a neural network to play the
game. This feature should have been developed on a separate branch but I
did not know enough about git to do that at the time. If I resume
development of it that is what I shall do, but this is low priority.