Changed name
I changed the name of the game to 'Fuel Leak!'. I think this is a much
better name than 'Asteroid Clone' but I am not entirely happy with it. I
have changed the splash screen, help menu, window name and README to
reflect the new name. I also want to change the repository name on
github, but this may be a larger issue.